+++ integrations/bultex_b2b/__init__.py
@@ -1,3 +1,4 @@
 from .client import BultexB2BClient
 from .parser import parse_product_page
 from .live_readonly import BultexReadOnlyClient, LoginFieldMap
+from .login_diagnostics import diagnose_login_page, LoginPageDiagnostics